requirements.txt
streamlit==1.38.0
python-dotenv==1.0.1
streamlit-authenticator==0.4.2
pyyaml==6.0.2